Compare bootstrapped distributions

Not usually called directly. Included in boot_predict. Usually used in combination with A function that takes the output from summary_factorlist(..., fit_id=TRUE) and merges with any number of model dataframes, usually produced with a model wrapper followed by the fit2df() function (see examples).

boot_compare(
  bs.out,
  confint_level = 0.95,
  confint_sep = " to ",
  comparison = "difference",
  condense = TRUE,
  compare_name = NULL,
  digits = c(2, 3),
  ref_symbol = 1
)

Arguments

bs.out

Output from boot::boot,

confint_level

The confidence level to use for the confidence interval. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

confint_sep

String separating lower and upper confidence interval

comparison

Either "difference" or "ratio".

condense

Logical. FALSE gives numeric values, usually for plotting. TRUE gives table for final output.

compare_name

Name to be given to comparison metric.

digits

Rounding for estimate values and p-values, default c(2,3).

ref_symbol

Reference level symbol

Value

A dataframe of first differences or ratios for boostrapped distributions of a metric of interest.

finalfit predict functions

See also

Examples

# See boot_predict.