finalfit: Quickly create elegant final results tables and plots when modelling.

finalfit: Quickly create elegant final results tables and plots when modelling.

finalfit model wrappers

glmuni, glmmulti, glmmulti_boot, glmmixed, lmuni, lmmulti, lmmixed, coxphuni, coxphmulti.

finalfit model extractor

Generic: fit2df

Methods (not called directly): fit2df.glm, fit2df.glmlist, fit2df.glmboot, fit2df.lm, fit2df.lmlist, fit2df.glmerMod, fit2df.lmerMod, fit2df.coxph, fit2df.coxphlist, fit2df.stanfit.

finalfit all-in-one function

Generic: finalfit.

Methods (not called directly): finalfit.glm, finalfit.lm, finalfit.coxph.

finalfit plotting functions

or_plot, hr_plot, surv_plot.

finalfit helper functions

ff_merge, ff_glimpse,

finalfit prediction functions

boot_predict, finalfit_newdata.

Methods (not called directly): boot_compare.

finalfit missing data functions

missing_pattern, missing_compare, missing_plot.