Missing values pairs plot

Compare the occurence of missing values in all variables by each other. Suggest limit the number of variables to a maximum of around six. Dependent and explanatory are for convenience of variable selection, are optional, and have no other specific function.

missing_pairs(
  .data,
  dependent = NULL,
  explanatory = NULL,
  use_labels = TRUE,
  title = NULL,
  position = "stack",
  showXAxisPlotLabels = TRUE,
  showYAxisPlotLabels = FALSE
)

Arguments

.data
Data frame.

dependent
Character vector. Optional name of dependent variable.

explanatory
Character vector. Optional name(s) of explanatory variables.

use_labels
Use variable label names in plot labelling.

title
Character vector. Optional title for plot.

position
For discrete variables, choose "stack" or "fill" to show counts or proportions.

showXAxisPlotLabels
Show x-axis plot labels.

showYAxisPlotLabels
Show y-axis plot labels.

Value

A plot matrix comparing missing values in all variables against each other.

Examples

if (FALSE) {
explanatory = c("age", "nodes", "age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = 'mort_5yr'
colon_s %>%
  missing_pairs(dependent, explanatory)
}