Western Collaborative Group Study

3154 healthy young men aged 39-59 from the San Francisco area were assessed for their personality type. All were free from coronary heart disease at the start of the research. Eight and a half years later change in this situation was recorded.

data(wcgs)

Format

A data frame with 3154 observations on the following 13 variables.

id Subject ID

age Age: age in years

height Height: height in inches

weight Weight: weight in pounds

sbp Systolic blood pressure: mmHg

dbp Diastolic blood pressure: mmHg

chol Cholesterol: mg/100 ml

personalityPersonality type/Behavior pattern: a factor with levels A1, A2, B3, B4

personality_2L Dichotomous personality type / behavior pattern: A = aggressive; B = passive

ncigs0 Smoking: Cigarettes/day

smoking Smoking: No, Yes

arcus0 Corneal arcus: No, Yes

chd Coronary heart disease event: No Yes

typechd coronary heart disease is a factor with levels No, MI_SD (MI or sudden death), Silent_MI, Angina

timechd Observation (follow up) time: Days

Source

Statistics for Epidemiology by N. Jewell (2004)

Details

The WCGS began in 1960 with 3,524 male volunteers who were employed by 11 California companies. Subjects were 39 to 59 years old and free of heart disease as determined by electrocardiogram. After the initial screening, the study population dropped to 3,154 and the number of companies to 10 because of various exclusions. The cohort comprised both blue- and white-collar employees. At baseline the following information was collected: socio-demographic including age, education, marital status, income, occupation; physical and physiological including height, weight, blood pressure, electrocardiogram, and corneal arcus; biochemical including cholesterol and lipoprotein fractions; medical and family history and use of medications; behavioral data including Type A interview, smoking, exercise, and alcohol use. Later surveys added data on anthropometry, triglycerides, Jenkins Activity Survey, and caffeine use. Average follow-up continued for 8.5 years with repeat examinations

References

Coronary Heart Disease in the Western Collaborative Group Study Final Follow-up Experience of 8 1/2 Years Ray H. Rosenman, MD; Richard J. Brand, PhD; C. David Jenkins, PhD; Meyer Friedman, MD; Reuben Straus, MD; Moses Wurm, MD JAMA. 1975;233(8):872-877. doi:10.1001/jama.1975.03260080034016.