finalfit: Quickly create elegant final results tables and plots when modelling.

Quickly create elegant final results tables and plots when modelling.

finalfit model wrappers

glmuni, glmmulti, glmmulti_boot, glmmixed, lmuni, lmmulti, lmmixed, coxphuni, coxphmulti, crruni, crrmulti, svyglmuni, svyglmmulti.

finalfit model extractor

Generic: fit2df

Methods (not called directly): fit2df.glm, fit2df.glmlist, fit2df.glmboot, fit2df.lm, fit2df.lmlist, fit2df.glmerMod, fit2df.lmerMod, fit2df.coxph, fit2df.coxphlist, fit2df.crr, fit2df.crrlist, fit2df.stanfit.

finalfit all-in-one function

Generic: finalfit. finalfit_permute.

Methods (not called directly): finalfit.glm, finalfit.lm, finalfit.coxph.

finalfit plotting functions

coefficient_plot, or_plot, hr_plot, surv_plot, ff_plot.

finalfit helper functions

ff_glimpse, ff_label, ff_merge, ff_interaction.

finalfit prediction functions

boot_predict, finalfit_newdata.

Methods (not called directly): boot_compare.

finalfit missing data functions

missing_glimpse, missing_pattern, missing_compare, missing_plot, missing_pairs.